+++ src/app/ui/layout-no-controls-unless-expanded.tsx
@@ -7,12 +7,16 @@ export class FullLayoutNoControlsUnlessExpanded extends Layout {
     override componentDidMount(): void {
         super.componentDidMount();
         // Hide Toggle Controls button unless expanded:
-        this.subscribe(this.plugin.layout.events.updated, () => {
-            const isExpanded = this.plugin.layout.state.isExpanded;
-            const hideCanvasControls = PluginCustomState(this.plugin).initParams?.hideCanvasControls;
-            const hideControlToggle = hideCanvasControls?.includes('controlToggle') || hideCanvasControls?.includes('all');
-            this.plugin.config.set(PluginConfig.Viewport.ShowControls, isExpanded && !hideControlToggle);
-        });
+        this.updateControlToggleVisibility();
+        this.subscribe(this.plugin.layout.events.updated, () => this.updateControlToggleVisibility());
+    }
+
+    /** Show/hide wrench button as needed (so it never displays in non-expanded mode) */
+    private updateControlToggleVisibility() {
+        const isExpanded = this.plugin.layout.state.isExpanded;
+        const hideCanvasControls = PluginCustomState(this.plugin).initParams?.hideCanvasControls;
+        const hideControlToggle = hideCanvasControls?.includes('controlToggle') || hideCanvasControls?.includes('all');
+        this.plugin.config.set(PluginConfig.Viewport.ShowControls, isExpanded && !hideControlToggle);
     }
 
     override get layoutVisibilityClassName(): string {